refactor: reorganized folder structure
