--- .env
@@ -1,5 +1,0 @@
-SECRET = '5a!85$4eRFKRghqX'
-DB_HOST = localhost
-DB = logsys
-DB_USER = root
-DB_PASS = ''
