--- História 2/Back/database/database.js
@@ -1,19 +1,0 @@
-const pg = require("pg");
-const Sequelize = require("sequelize");
-const sequelize = new Sequelize("users", "root", "postgres", {
-  dialect: "pg",
-  host: "PG_HOST",
-  port: "PG_PORT",
-});
-
-const dotenv = require("dotenv");
-
-dotenv.config();
-
-const Pool = pg.Pool;
-
-const config = `postgresql://${process.env.PG_USERNAME}:${process.env.PG_PASSWORD}@${process.env.PG_HOST}:${process.env.PG_PORT}/${process.env.PG_DATABASE}`;
-
-const pool = new Pool({ connectionString: config });
-
-(module.exports = pool), (module.exports = sequelize);
