--- História 2/Back/produto.js
@@ -1,11 +1,0 @@
-const Sequelize = require("sequelize");
-const database = require("../database/database");
-
-const users = database.define("users", {
-  nome: {
-    type: Sequelize.INTEGER,
-    autoIncrement: true,
-    allowNull: false,
-    primaryKey: true,
-  },
-});
