--- models/usuarios.js
@@ -1,23 +1,0 @@
-const Sequelize = require('sequelize');
-const db = require('./db');
-
-const usuarios = db.define('users',{
-    id:{
-        type: Sequelize.INTEGER,
-        autoIncrement: true,
-        allowNull: false,
-        primaryKey: true
-    },
-    email: {
-        type: Sequelize.STRING,
-        allowNull: false,        
-    },
-    password: {
-        type: Sequelize.STRING,
-        allowNull: false,
-    }
-});
-
-//usuarios.sync();
-
-module.exports = usuarios;
