--- package.json
@@ -1,28 +1,0 @@
-{
-  "name": "back",
-  "version": "1.0.0",
-  "description": "backend para página de lpgin",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "github"
-  },
-  "keywords": [
-    "login"
-  ],
-  "author": "LilianDamico",
-  "license": "ISC",
-  "dependencies": {
-    "bcryptjs": "^2.4.3",
-    "body-parser": "^1.20.2",
-    "cors": "^2.8.5",
-    "dotenv": "^16.0.3",
-    "express": "^4.18.2",
-    "jsonwebtoken": "^9.0.0",
-    "mysql2": "^3.3.1",
-    "sequelize": "^6.31.1"
-  }
-}
